add service payment

+++ EcommerWebCar/main-service/src/main/java/com/tlaq/main_service/dto/requests/carRequest/CarServiceRequest.java
@@ -3,6 +3,8 @@ package com.tlaq.main_service.dto.requests.carRequest;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,13 +13,13 @@ import lombok.experimental.FieldDefaults;
 public class CarServiceRequest {
     String dongCo;
     String hopSo;
-    Double congSuat;
-    Double momenXoan;
-    Double dungTichXiLanh;
-    Double dungTichXang;
-    Double taiTrong;
-    Double chieuDai;
+    BigDecimal congSuat;
+    BigDecimal momenXoan;
+    BigDecimal dungTichXiLanh;
+    BigDecimal dungTichXang;
+    BigDecimal taiTrong;
+    BigDecimal chieuDai;
     String mauSac;
-    Double tocDoToiDa;
+    BigDecimal tocDoToiDa;
     String loaiNhienLieu;
 }
